requirements.txt
streamlit
openai
PyYAML
pandas